improvement(client): add duplicate Signal telemetry
Introduce a missing Signal count spanning entire
connection lifetime. And a lower bound for most
recent sequence of Signals that have been received
without and missing Signals.

Use those details to collect when a Signal must be
a duplicate of an already received Signal rather
than an out-of-order Signal.

Some duplicate Signals of older Signals may still
be reported as out-of-order Signals.